+++ plot/rgb_image.go
@@ -1,6 +1,7 @@
 package plot
 
 import (
+	"fmt"
 	"image/color"
 
 	"github.com/faiface/pixel"
@@ -16,37 +17,31 @@ import (
 // The image is scaled to the canvas extent, with preserved aspect ratio.
 // Does not add plot axes etc.
 type ImageRGB struct {
-	Scale     float64           // Spatial scaling: cell size in screen pixels. Optional, default auto.
-	Observers []observer.Matrix // Observers for the red, green and blue channel. Elements can be nil.
-	Min       []float64         // Minimum value for channel color mapping. Optional, default [0, 0, 0].
-	Max       []float64         // Maximum value for channel color mapping. Optional, default [1, 1, 1].
-	slope     []float64
-	dataLen   int
-	picture   *pixel.PictureData
+	Scale    float64               // Spatial scaling: cell size in screen pixels. Optional, default auto.
+	Observer observer.MatrixLayers // Observer providing data for color channels.
+	Layers   []int                 // Layer indices. Optional, defaults to [0, 1, 2].
+	Min      []float64             // Minimum value for channel color mapping. Optional, default [0, 0, 0].
+	Max      []float64             // Maximum value for channel color mapping. Optional, default [1, 1, 1].
+	slope    []float64
+	dataLen  int
+	picture  *pixel.PictureData
 }
 
 // Initialize the drawer.
 func (i *ImageRGB) Initialize(w *ecs.World, win *pixelgl.Window) {
-	if len(i.Observers) != 3 {
-		panic("RgbImage plot needs exactly 3 observers")
+	i.Observer.Initialize(w)
+
+	if i.Layers == nil {
+		i.Layers = []int{0, 1, 2}
+	} else if len(i.Layers) != 3 {
+		panic("rgb image plot Layers must be of length 3")
 	}
-	width, height := -1, -1
-	for j := 0; j < 3; j++ {
-		if i.Observers[j] == nil {
-			continue
+
+	layers := i.Observer.Layers()
+	for _, l := range i.Layers {
+		if layers <= l {
+			panic(fmt.Sprintf("layer index %d out of range", l))
 		}
-		i.Observers[j].Initialize(w)
-		wi, he := i.Observers[j].Dims()
-		if width >= 0 && width != wi {
-			panic("observers differ in matrix width")
-		}
-		if height >= 0 && height != he {
-			panic("observers differ in matrix width")
-		}
-		width, height = wi, he
-	}
-	if width < 0 && height < 0 {
-		panic("needs an observer for at least one channel")
 	}
 
 	if i.Min == nil {
@@ -68,17 +63,14 @@ func (i *ImageRGB) Initialize(w *ecs.World, win *pixelgl.Window) {
 		1.0 / (i.Max[2] - i.Min[2]),
 	}
 
+	width, height := i.Observer.Dims()
 	i.dataLen = width * height
 	i.picture = pixel.MakePictureData(pixel.R(0, 0, float64(width), float64(height)))
 }
 
 // Update the drawer.
 func (i *ImageRGB) Update(w *ecs.World) {
-	for j := 0; j < 3; j++ {
-		if i.Observers[j] != nil {
-			i.Observers[j].Update(w)
-		}
-	}
+	i.Observer.Update(w)
 }
 
 // UpdateInputs handles input events of the previous frame update.
@@ -86,19 +78,12 @@ func (i *ImageRGB) UpdateInputs(w *ecs.World, win *pixelgl.Window) {}
 
 // Draw the drawer.
 func (i *ImageRGB) Draw(w *ecs.World, win *pixelgl.Window) {
-	cannels := make([][]float64, 3)
-	for j := 0; j < 3; j++ {
-		if i.Observers[j] != nil {
-			cannels[j] = i.Observers[j].Values(w)
-		}
-	}
+	cannels := i.Observer.Values(w)
 
 	values := append([]float64{}, i.Min...)
 	for j := 0; j < i.dataLen; j++ {
-		for k := 0; k < 3; k++ {
-			if cannels[k] != nil {
-				values[k] = cannels[k][j]
-			}
+		for i, k := range i.Layers {
+			values[i] = cannels[k][j]
 		}
 		i.picture.Pix[j] = i.valuesToColor(values[0], values[1], values[2])
 	}